Temp Art added, refined word tossing animation
Falling word gets tossed left at the monster. Temp images are
positioned relative to keyboard to optimize placement of assets for
iPhone 4S & above.

User Input word correctly blanks itself upon completion of the typed
word, but I need to implement it so that this behavior happens when the
user presses return.